requirements.txt
flask
google-api-python-client
openai
python-dotenv
google.generativeai
requests
bs4
nltk
scikit-learn
transformers
torch
sentence-transformers
faiss-cpu